services/github/requirements.txt
Flask==1.0.2
python-http-client==3.1.0
pytest==4.6.4
gunicorn==19.9.0
github3.py==0.9.5
